backend/pyproject.toml
[project]
name = "hostech-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "prawinsanjai",email = "prawinpk10@gmail.com"}
]
requires-python = ">=3.11,<4"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy-utils (>=0.42.1,<0.43.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "django-environ (>=0.12.1,<0.13.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
